+++ packages/entities/entities-plugins/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@2.4.1...@kong-ui-public/entities-plugins@2.4.2) (2023-12-01)
+
+
+### Bug Fixes
+
+* delete modal ([#980](https://github.com/Kong/public-ui-components/issues/980)) ([5a67bca](https://github.com/Kong/public-ui-components/commit/5a67bcaddecfaf79a0bb3db37e160a562367e9b7))
+
+
+
+
+
+## [2.4.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@2.4.0...@kong-ui-public/entities-plugins@2.4.1) (2023-12-01)
+
+**Note:** Version bump only for package @kong-ui-public/entities-plugins
+
+
+
+
+
+# [2.4.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@2.3.3...@kong-ui-public/entities-plugins@2.4.0) (2023-12-01)
+
+
+### Features
+
+* **entities-plugins:** endpoint fixes ([#976](https://github.com/Kong/public-ui-components/issues/976)) ([17cdc92](https://github.com/Kong/public-ui-components/commit/17cdc92a70612956aac97a9fb9ff97e1e9b90a4c))
+
+
+
+
+
 ## [2.3.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@2.3.2...@kong-ui-public/entities-plugins@2.3.3) (2023-12-01)